Reduce number of units converted without explicitly asking
Set only a small nubmer of units to automatically convert without any of
the indicators that the user wants a more comprehensive set of
conversion results. For temperature, for example, automatically convert
when seeing a fahrenheit or celsius value, but not kelvin. This should
cut down on unwanted spam from units that may also be common terms
elsewhere (like fps, which can mean feet per second or frames per second
in more common usage).